Moving more functions out of lib
Beginning the process of rewriting CFFT to use RowMajorMatrix instead of vec.

Currently completely broken.
