--- cfft/src/naive.rs
@@ -1,4 +1,0 @@
-/// A Naive implementation of the CFT computing evaluations/coefficients directly.
-
-#[derive(Default, Clone)]
-pub struct NaiveCft;
